Move gw types to response.rs
